backend/requirements.txt
# Core Django
Django>=4.2,<5.0
psycopg2-binary>=2.9.5

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-filter>=23.0.0

# CORS handling for frontend integration
django-cors-headers>=4.0.0

# Environment variables
python-decouple>=3.8
python-dotenv>=1.0.0

# Database
dj-database-url>=2.0.0

# Authentication & Security
django-allauth>=0.54.0
django-oauth-toolkit>=2.2.0
cryptography>=41.0.0

# API Documentation
drf-spectacular>=0.26.0
drf-yasg>=1.21.0

# File handling & Storage
Pillow>=10.0.0
django-storages>=1.13.0
boto3>=1.28.0  # For AWS S3 storage if needed

# Caching
redis>=4.5.0
django-redis>=5.2.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0  # For HTML parsing
celery>=5.3.0  # For background tasks
python-dateutil>=2.8.0

# AI Services
openai>=1.3.0
anthropic>=0.7.0

# Image Management
cloudinary>=1.36.0

# WebSocket Support
channels>=4.0.0
channels-redis>=4.1.0
pytz>=2023.3
playwright>=1.40.0  # For browser automation

# Development & Testing
django-debug-toolbar>=4.1.0
factory-boy>=3.2.0
pytest>=7.4.0
pytest-django>=4.5.0
pytest-asyncio>=0.23.5
coverage>=7.2.0

# Linting & Code Quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Production
gunicorn>=21.0.0
whitenoise>=6.5.0  # For serving static files

# Monitoring & Logging
sentry-sdk>=1.28.0
django-extensions>=3.2.0
